+++ ddcloud/provider.go
@@ -111,6 +111,15 @@ func Provider() terraform.ResourceProvider {
 			// A virtual listener is the top-level entity for load-balancing functionality.
 			"ddcloud_virtual_listener": resourceVirtualListener(),
 
+			// An SSL-offload profile for a virtual listener.
+			"ddcloud_ssl_offload_profile": resourceSSLOffloadProfile(),
+
+			// An SSL certificate (with private key) for a domain.
+			"ddcloud_ssl_domain_certificate": resourceSSLDomainCertificate(),
+
+			// An SSL certificate chain.
+			"ddcloud_ssl_certificate_chain": resourceSSLCertificateChain(),
+
 			// A reserved IPv6 or private IPv4 address on a VLAN.
 			"ddcloud_ip_address_reservation": resourceIPAddressReservation(),
 		},
@@ -121,6 +130,9 @@ func Provider() terraform.ResourceProvider {
 
 			// A virtual network (VLAN).
 			"ddcloud_vlan": dataSourceVLAN(),
+
+			// A PKCS12 (PFX) file.
+			"ddcloud_pfx": dataSourcePFX(),
 		},
 
 		// Provider configuration